+++ web/app/connectors/_merlins-connector/app/commands.js
@@ -9,7 +9,16 @@ import {browserHistory} from 'progressive-web-sdk/dist/routing'
 import {parseLoginStatus, parseSearchSuggestions} from './parser'
 import {parseNavigation} from '../navigation/parser'
 import {receiveFormKey} from '../actions'
-import {CHECKOUT_SHIPPING_URL, WISHLIST_URL, SIGN_IN_URL, CART_URL, buildQueryURL, buildSearchURL} from '../config'
+import {
+    CHECKOUT_SHIPPING_URL,
+    WISHLIST_URL,
+    SIGN_IN_URL,
+    CART_URL,
+    ACCOUNT_ADDRESS_URL,
+    ACCOUNT_INFO_URL,
+    buildQueryURL,
+    buildSearchURL
+} from '../config'
 import {getCookieValue} from '../../../utils/utils'
 import {generateFormKeyCookie} from '../../../utils/magento-utils'
 import {setPageFetchError} from 'progressive-web-sdk/dist/store/offline/actions'
@@ -22,7 +31,9 @@ import {
     setCartURL,
     setWishlistURL,
     setLoggedIn,
-    setSignInURL
+    setSignInURL,
+    setAccountAddressURL,
+    setAccountInfoURL
 } from 'progressive-web-sdk/dist/integration-manager/results'
 
 const requestCapturedDoc = () => {
@@ -85,6 +96,8 @@ export const initApp = () => (dispatch) => {
     const formKey = getCookieValue('form_key') || generateFormKeyCookie()
     dispatch(receiveFormKey(formKey))
 
+    dispatch(setAccountAddressURL(ACCOUNT_ADDRESS_URL))
+    dispatch(setAccountInfoURL(ACCOUNT_INFO_URL))
     dispatch(setCheckoutShippingURL(CHECKOUT_SHIPPING_URL))
     dispatch(setWishlistURL(WISHLIST_URL))
     dispatch(setSignInURL(SIGN_IN_URL))