+++ web/app/containers/reducers.js
@@ -7,6 +7,7 @@
 import {combineReducers} from 'redux'
 
 import navigation from '../modals/navigation/reducer' // navigation is a special case, since it's not in the containers directory
+import accountAddress from './account-address/reducer'
 import accountDashboard from './account-dashboard/reducer'
 import app from './app/reducer'
 import cart from './cart/reducer'
@@ -23,6 +24,7 @@ import productList from './product-list/reducer'
 
 const uiReducer = combineReducers({
     navigation,
+    accountAddress,
     accountDashboard,
     app,
     cart,