+++ web/CHANGELOG.md
@@ -1,6 +1,9 @@
 ## To be released
 - Bump Lighthouse to version 2.2.1 [889](https://github.com/mobify/platform-scaffold/pull/889)
 - Add product variation support to Merlin's connector [#800](https://github.com/mobify/platform-scaffold/pull/800)
+- Add account pages
+    - Account Dashboard[#834](https://github.com/mobify/platform-scaffold/pull/834)
+    - Account Addresses[#886](https://github.com/mobify/platform-scaffold/pull/886)
 
 ## 0.17.3 (July 7, 2017)
 - Moved a pair of browser storage detection methods to SDK [#850](https://github.com/mobify/platform-scaffold/pull/850)